requirements/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary
dj-database-url
python-decouple
phonenumbers
django-filter
drf-yasg
django-cors-headers
djangorestframework-simplejwt